+multiple user support complete
Replaced get_all endpoing with has_session, start_session.

has_session: relays account data if there is a session cookie (with an
'email' value), allowing it to redirect. Otherwise, it sits on the
contact email page.

start_session: Returns the data associated with the submitted email. If
the email has no entry in the contact_details table, a new one is
created and the default data is returned.