+bugfix: Email update
Prior restructuring of the DBs, specifically removing the dedicated 'id'
columns designated as the primary keys and instead using the email field
was a mistake. Using that field as a foreign key means it cannot be
updated when foreign key constraints are enforced.  Fixed by allowing a
call to get_db without enforcing foreign keys.